refactor: extract match title requirements
